% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7D10.R
\name{C7D10}
\alias{C7D10}
\title{A data generating model with Seven clusters}
\usage{
C7D10(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Seven clusters in multiple dimensions having unequal within cluster variations. All the clusters are generated from Gaussain distributions.
}
\details{
The data set has ten dimensions.
}
\examples{
dmat <- C7D10(700)
plot2d(dmat$data, dmat$truelab)
pairplots(dmat$data, dmat$truelab) 
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
