% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCo4D2.R
\name{CCo4D2}
\alias{CCo4D2}
\title{A data generating model with four clusters}
\usage{
CCo4D2(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Generates a data set consists of 4-clusters, one each from the specified Gaussian distributions.
}
\details{
The data set has two iid dimensions and \code{K=4} clusters.
}
\examples{
dmat <- CCo4D2(1000)
plot2d(dmat$data, dmat$truelab)
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
