% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElongUniGaussian.R
\name{ElongUniGaussian}
\alias{ElongUniGaussian}
\title{A data generating model with 9-clusters}
\usage{
ElongUniGaussian(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Generates data set consists of 9-clusters, from Gaussian and Uniform distributions. See details for complete model definition.
}
\details{
The data set has two dimensions.
}
\examples{
dmat <- ElongUniGaussian(900)
plot2d(dmat$data, dmat$truelab)

}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
