% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasesUniGaussian.R
\name{FarUniGauss}
\alias{FarUniGauss}
\title{A data generating model with two clusters}
\usage{
FarUniGauss(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Generates a data set consists of 2-clusters, one each from the Gaussian and Uniform distributions.
}
\details{
The data set has two iid dimensions.
}
\examples{
dmat <- FarUniGauss(1000)
plot2d(dmat$data, dmat$truelab)
plot(dmat$data, col=dmat$truelab, xlim=c(-20, 20), ylim=c(-20, 20))

}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
