% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosilFix.R
\name{fosilFix}
\alias{fosilFix}
\title{Fast osil-fixed number of clusters}
\usage{
fosilFix(dmat, distmethod="euclidean", k)
}
\arguments{
\item{dmat}{Either a numeric matrix or data frame of observed values.
The row represent observations to cluster and columnn represents the variables.}

\item{distmethod}{the distance method to be used. Current available methods are "euclidean", 
"maximum", "manhattan", "canberra", "binary" or "minkowski". See \code{\link[stats:dist]{dist}} for more details on these methods.}

\item{k}{number of clusters}
}
\value{
Returns a list having following components:
\describe{
\item{n}{total number of data points.}
\item{K}{estimated number of clusters.}
\item{clus_lab}{clustering label vector.}
\item{clus_size}{cluster sizes.}
\item{silh}{OASW value of each cluster.}
\item{avg_clus_silh}{OASW for each cluster.}
\item{avg_silh}{OASW value for the clustering.}
\item{$avg_silh_kmin_kmax}{ASW value for all the clusterings.}
}
}
\description{
This is the fast version of the OSil algorithm for a single number of cluster. 
OSil is an optimum average silhouette width (OASW) clustering method that 
donot use any kind of cluster centriods. Only data is needed as input. 
The algorithm can estimate number of clusters.
}
\details{
osil has an initialization phase. Based on extensive simulaitons the best
 initialization methods from among a wide range of existing clustering methods, 
  for the algorithm has been identified namely average, Ward's, pam,
   kmeans, and model-based clustering. In case distances are provided for 
   clustering kmeans and model-based clustering are excluded from the initialization methods.
}
\examples{
require(mlbench)
dmat <- mlbench.shapes(100)$x
oasw_clus <- fosil(dmat) 
k <- 4
oasw_clus <- fosilFix(dmat, distmethod="euclidean", k)
plot(dmat, col = oasw_clus$clus_lab, pch = 16, cex = 1.5)
}
\references{
Batool F. (2019). Optimum average silhouette width clustering. 
\emph{PhD Thesis}, University College London.
}
\seealso{
\code{\link{osil}} for not so fast version.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
