% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosil.R
\name{hosil}
\alias{hosil}
\title{Hierarchical optimum average silhouette width clustering}
\usage{
hosil(dys,  distmethod = "euclidean", fixK = "NA")
}
\arguments{
\item{dys}{A vector of pairwise distances between observations. Usually an object of class 
\code{"dist"} or a data matrix or data frame. In latter case also needs distance method. 
The default is set at \code{euclidean}.}

\item{distmethod}{the distance mehtod to be used. Current available methods are
\code{euclidean}, \code{maximum}, \code{manhattan}, \code{canberra}, \code{binary}, or 
\code{minkowski}. See \code{\link[stats:dist]{dist}} for more details on these methods.}

\item{fixK}{user defined number of clusters against which clustering is requried.}
}
\value{
Returns a list with the follwoing components
\describe{
\item{est_K}{estimated number of clusters.}
\item{OASW_est_K}{Value of the objective function against the estimated number of clusters.}
\item{clus_vect_est}{clustering label vector for the estimated number of clusters.}
\item{all_OASW}{Values of the objective function (OASW linkage) against \code{n-1} to \code{2} 
number of clusters. Thus the first value \code{all_OASW[1]} gives the OASW linkage value 
against \code{(n-1)} number of clusters, the next value contains \code{(n-2)} 
then \code{(n-3),..., 3, 2}.}

\item{fix_K}{user specified number of clusters }
\item{clus_vect_fix}{clustering label vector for the user specified number of clusters}
\item{OASW_fix_K}{Value of the objective function against the user specified number of clusters}
}
}
\description{
hosil is a clustering alogromative hierarchical clustering algorithm.
 The cluster mearge are defined at each hierarchy using a new linkage method defined by
  the optimization of the ASW index. The method can also estimate the number of clusters 
  based on OASW linkage criterion. If number of clsuters is to be fixed
   (see \code{fixK}) the minimum allowed is 3 and can be at most \code{(n-1)}.
}
\examples{
dmat <- MultiDist(350)
dys <- dist(dmat$data)
oasw_clustering <- hosil(dys) 
oasw_clustering <- hosil(dys, fixK=6)
plot(dmat, col = oasw_clus$clus_vect_fix, pch = 16)
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.

Kaufman, L. and P. J. Rousseeuw (1990). Finding groups in data: an introduction to cluster analysis,
 Volume 344. John Wiley & Sons.
}
\author{
Fatima Batool \email{ ucakfba@ucl.ac.uk }
}
