% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{initialization function for OASW clustering algorithms}
\usage{
init(dmat, K, distmethod = "euclidean", ...)
}
\arguments{
\item{dmat}{either a numeric matrix or data frame of observed values or 
pairwise distances between observations.}

\item{K}{number of cluster}

\item{distmethod}{distance method to be used for the calculation of
pairwise distances between observations.}

\item{...}{additional parameters}
}
\value{
Returns a list having following components
\describe{
\item{n}{number of observations.}
\item{K}{number of clusters.}
\item{lab_best}{clustering labels corresponding to the best ASW clustering among the  initialization methods. }
\item{asw_best }{best ASW value among the initialization methods. }
\item{best_init_method}{name of the best initialization methods based on ASW value. }
}
}
\description{
This function was originally wrote to provide the initialization for osil
 and fosil but can have standalone usage.  The function takes a fixed number of clusters.
}
\details{
The function is originally written as an initialization function for osil and fosil 
clusterings. The \code{init} functions returns the best clustering out of the six 
methods initialization methods based on the ASW values. Several clustering methods were considered
in a systematic simulation set-up to find out the best for the OASW clustering initialization.
 Among those considered six showed good performance for single or multiple data generating structures 
 considered in the simulations (see Batool, 2019 for results and discussion).
The six best initialization methods are average (Sokal and Michener 1958), Ward's  (Ward 1963), pam (Kaufman and Rousseeuw 1990), 
kmeans (Hartigan and Wong 1979), and model-based (Fraley and Raftery 1998) clustering. Works with both distances and data matrix.
Currently, in case distances are provided \code{kmeans} and \code{Mclust} are
excluded from initialization methods.
}
\examples{
dmat <- TwoGaussian(100)
dys <- dist(dmat$data)
plot(dmat$data, col = dmat$truelab)
init_res_1 <- init(dmat$data, KK=2, distmethod = "euclidean")
init_res_2 <- init(dys, KK=2, distmethod = "euclidean")
print(init_res_1)
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.

C. Fraley and A. E. Raftery (1998). How many clusters? which clustering method? 
answers via model-based cluster analysis. \emph{The Computer Journal}, 41(8):578 588.

Hartigan, J. A. and Wong, M. A. (1979). Algorithm AS 136: 
A K-means clustering algorithm. \emph{Applied Statistics}, 28, 100 108.

J. H. Ward Jr. (1963). Hierarchical grouping to optimize an objective function. 
\emph{Journal of the American Statistical Association}, 58(301):236 244.

Kaufman, L. and P. J. Rousseeuw (1990). Finding groups in data: 
an introduction to cluster analysis, Volume 344. John Wiley & Sons.

McQuitty, L. L. (1957). Elementary linkage analysis for isolating orthogonal and oblique 
types and typal relevancies. Educational and PsychologicalMeasurement 17(2), 207 229.
 
R. Sokal and C. D. Michener (1958). A statistical method for evaluating systematic 
relationships. \emph{Univesity Kansas Science Bulletin}, 38(22):1409 1438.
}
\seealso{
\code{\link{hclust}}, \code{\link{pam}}, \code{\link{kmeans}}, \code{\link{Mclust}}
functions to pass additional argumnets to \code{\link{init}}.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
