% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoisyGaussian.R
\name{NoisyGaussian}
\alias{NoisyGaussian}
\title{A data generating model with two clusters and added Uniform noise points}
\usage{
NoisyGaussian(n,noise.points)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}

\item{noise.points}{number of noie points requried to be added in data set}
}
\value{
Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Generates data set consists of two clossely located Gaussian clusters with 500 Uniform noise points.
}
\details{
The data set has two iid dimensions.
}
\examples{
dmat <- NoisyGaussian(200, 50)
plot(dmat$data, col=c("blue","green", "red")[dmat$truelab])
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
