% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osilFix.R
\encoding{UTF-8}
\name{osilFix}
\alias{osilFix}
\title{osil clustering-fixed number of clusters}
\usage{
osilFix(dys, n, K, clus_lab)
}
\arguments{
\item{dys}{pairwise distances between observations.}

\item{n}{number of observations.}

\item{K}{number of clusters.}

\item{clus_lab}{clustering labels.}
}
\value{
Returns a list having following components.
\describe{
\item{n}{total number of data points.}
\item{K}{estimated number of clusters.}
\item{clus_lab}{clustering labels.}
\item{clus_size}{cluster sizes.}
\item{silh}{ASW value of each cluster.}
\item{avg_clus_silh}{ASW for each cluster.}
\item{avg_silh}{ASW value for the clustering.}
\item{iter}{number of iteration taken by the algorithm to converge.} }
}
\description{
Produces a clustering solution for a fixed number of clusters
}
\details{
This is a wrapper function for C++ functions. 
This function is called from within osil function.  
Can be called standalone for clustering for fixed 
number of clusters with an intialization clustering label set. 
This will return an OASW clustering based on labels.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
