% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplots.R
\name{pairplots}
\alias{pairplots}
\title{pairplots}
\usage{
pairplots(data, labels)
}
\arguments{
\item{data}{data set to plot.}

\item{labels}{labels against which plotting is needed.}
}
\value{
Returns a plot.
}
\description{
pair plots for more than two demensional data
}
\details{
none
}
\examples{
dmat <- C7D10(70)
pairplots(dmat$data, dmat$truelab)   
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
