% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamsil.R
\name{pamsil}
\alias{pamsil}
\title{pamsil clustering}
\usage{
pamsil(dmat, distmethod = "euclidean", kmin=2, kmax=12)
}
\arguments{
\item{dmat}{Either a matrix or data frame of observed values or
a vector of pairwise distances between observations. 
In first case the row represent observations to cluster and
columnn represents the variables. If data matrix is provided needs 
to specify the distance method as well. In second case usually an
object of class \code{"dist"}. Missing values are not allowed in both cases.}

\item{distmethod}{distance method to be used.}

\item{kmin}{minimum number of clusters for estimation of number of clusters.}

\item{kmax}{maximum number of clusters for estimation of number of clusters.}
}
\value{
Returns a list having following components:
\describe{
\item{est_K}{ number of clusters estimated by pamsil algorithms. All other results are based on this estimated number.}
\item{clus_lab}{pamsil clustering labels against estimated \code{K}.}
\item{silh}{optimum ASW value for each data point in the clustering.}
\item{clus_size}{number of observations in clusters. }
\item{avg_silh}{ASW value for pamsil clustering.}
\item{avg_clus_silh}{ASW for each cluster.}
\item{iter}{number of iteration taken by the algorithm to converge.}
\item{avg_silh_kmin_kmax}{ASW values for the clusterings correspondng to the number of clusters in the range \code{kmin} to \code{kmax} number of clusters.}
}
}
\description{
An OASW clustering method based on medoids.
}
\details{
This function is based on the standalone \code{C} functions written by Van et al. (2003). 
\code{pamsil()} sccepts both data matrix and distances.
}
\examples{
dmat <- iris[,1:4]
dys <- dist(dmat)
oasw_clustering <- pamsil(dys, 2, 4)
oasw_clustering <- pamsil(dmat, distmethod = "manhattan", 2, 4)

}
\references{
Van der Laan, M., Pollard, K., & Bryan, J. (2003). A new partitioning around medoids
 algorithm. \emph{Journal of Statistical Computation and Simulation}, 73(8), 575-584.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
