% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamsilFix.R
\name{pamsilFix}
\alias{pamsilFix}
\title{A new partitioning around medoids algorithm}
\usage{
pamsilFix(dys, K, distmethod = "euclidean")
}
\arguments{
\item{dys}{Either a matrix or data frame of observed values or a vector of pairwise distances between observations. In first case the row represent observations to cluster and columnn represents the variables. If data matrix is provided needs to specify the distance method as well. In second case usually an object of class \code{"dist"}. Missing values are not allowed in both cases.}

\item{K}{number of clusters}

\item{distmethod}{distance method to be used}
}
\value{
Returns a list having following components:
\describe{
\item{clus_lab}{pamsil clustering labels.}
\item{silh}{ASW value for each data point in the clustering.}
\item{clus_size}{number of observations in clusters.}
\item{avg_clus_silh}{ASW for each cluster.}
\item{avg_silh}{ASW value for the clustering.}
\item{iter}{number of iteration taken by the algorithm to converge.}
}
}
\description{
An OASW clustering method based on medoids. 
A PAM like clustering algorithm for the optimization of ASW based on medoids proposed in Van et al. (2003).he
}
\details{
This function is based on the standalone \code{C} functions written by Van et al. (2003). Use of this 
function is recommended if number of clusters are known. \code{pamsilFix} accepts both data matrix and distances.
}
\examples{
dmat <- iris[,1:4]
dys <- dist(dmat)
oasw_clustering <- pamsilFix(dys, 3)
oasw_clustering <- pamsilFix(dmat, 3, distmethod = "manhattan")

}
\references{
Van der Laan, M., Pollard, K., & Bryan, J. (2003). A new partitioning around medoids
 algorithm. \emph{Journal of Statistical Computation and Simulation}, 73(8), 575-584.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
