% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2d.R
\name{plot2d}
\alias{plot2d}
\title{Enhanced 2d plotting}
\usage{
plot2d(data, labels)
}
\arguments{
\item{data}{data to plot}

\item{labels}{labels against which plotting is needed}
}
\value{
Returns a plot
}
\description{
plots for two demensional data
}
\details{
none
}
\examples{
dmat <- C7D10(1000)
plot2d(dmat$data, dmat$truelab) 
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
