% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeMicroarray.R
\name{ThreeMicroarray}
\alias{ThreeMicroarray}
\title{A data generating model with three microarray like clusters}
\usage{
ThreeMicroarray(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Generates data set consists of three clusters.
}
\details{
The data set has 1000-dimensions. The dimensions are iid.
}
\examples{
dmat <- ThreeMicroarray(120)
plot2d(dmat$data, dmat$truelab)

}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
