% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeMicroarrayMulti.R
\name{ThreeMicroarrayMulti}
\alias{ThreeMicroarrayMulti}
\title{A data generating model to simulate microarray-like settings}
\usage{
ThreeMicroarrayMulti(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Generates data set consists of three clusters as defined in Van der Laan (2003).
}
\details{
The data set has 1000-dimensions and \code{K=7} number of clusters. The dimensions are iid.
}
\examples{
dmat <- ThreeMicroarrayMulti(60)
}
\references{
Van der Laan, M., K. Pollard, and J. Bryan (2003). A new partitioning around medoids algorithm. 
\emph{Journal of Statistical Computation and Simulation}, 73(8), 575 584.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
