% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNestedGaussian.R
\name{TNestedGaussian}
\alias{TNestedGaussian}
\title{A data generating model with three nested clusters}
\usage{
TNestedGaussian(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Generates data set consists of three clusters. Two closely located Gaussain clusters are nested inside Student's t cluster.
}
\details{
The data set has two iid dimensions.
}
\examples{
dmat <- TNestedGaussian(300)
plot(dmat$data, col=c("blue", "red","green")[dmat$truelab], xlab = " ", ylab = " ")
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.

Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
