% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitGaussian3D.R
\name{UnitGaussian3D}
\alias{UnitGaussian3D}
\title{A data generating model with nine clusters in three dimensions}
\usage{
UnitGaussian3D(n)
}
\arguments{
\item{n}{number of observations in the dataset. These are equally divided between clusters.}
}
\value{
Returns a list having two components described as follows:

Returns a list having two components described as follows;
 \describe{
 \item{n}{number of observations.}
  \item{K}{number of clusters.}
   \item{cluster size}{number of observations in each cluster.}
 \item{data}{data set generated from the model.}
\item{truelab}{clustering label vector correspounding to the known data generating model.}
}
}
\description{
Generates data set consists of nine clusters each drawn from a Gaussian distribution. See details 
for complete discription of the model's structure.
}
\details{
The data set has three iid dimensions.
}
\examples{
dmat <- UnitGaussian3D(891)
plot2d(dmat$data, dmat$truelab)
scatter3D(data[,1], data[,2], data[,3], colkey = FALSE, colvar = truelab, 
 ticktype = "detailed", pch = 16, bty = "b")
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.  
\emph{PhD Thesis}, University College London.
}
\author{
Fatima Batool \email{ucakfba@ucl.ac.uk}
}
